+++ test/extract/module-options/Binders.hs
@@ -1,0 +1,5 @@
+module Binders where
+
+{-# ANN f "doctest-parallel: --no-randomize-error" #-}
+f :: a -> a
+f = id