+++ test/extract/module-options/Mono.hs
@@ -1,0 +1,8 @@
+module Mono where
+
+{-# ANN module "doctest-parallel: --no-randomize-error1" #-}
+{-# ANN module ("doctest-parallel: --no-randomize-error2") #-}
+{-# ANN module ("doctest-parallel: --no-randomize-error3"  ) #-}
+{-# ANN module ("doctest-parallel:   --no-randomize-error4"  ) #-}
+{-# ANN module ("doctest-parallel:   --no-randomize-error5   "  ) #-}
+{-# ANN module ("doctest-parallel: --no-randomize-error6" :: String) #-}