+++ test/extract/module-options/NoOptions.hs
@@ -1,0 +1,5 @@
+module NoOptions where
+
+{-# ANN module "doctest-parallel" #-}
+{-# ANN module "abc" #-}
+{-# ANN module "" #-}