+++ test/extract/module-options/Poly.hs
@@ -1,0 +1,5 @@
+{-# LANGUAGE OverloadedStrings #-}
+
+module Poly where
+
+{-# ANN module ("doctest-parallel: --no-randomize-error" :: String) #-}