+++ test/integration/ModuleOptions/Foo.hs
@@ -1,0 +1,22 @@
+module ModuleOptions.Foo where
+
+{-# ANN module "doctest-parallel: --preserve-it" #-}
+
+-- |
+--
+-- >>> :t 'a'
+-- 'a' :: Char
+--
+-- >>> "foo"
+-- "foo"
+--
+-- >>> length it
+-- 3
+--
+-- >>> it * it
+-- 9
+--
+-- >>> :t it
+-- it :: Int
+--
+foo = undefined